# src/orders-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orders API Documentation",
        "description": "Swagger API Documentation for Orders managament API",
        "contact": {
            "email": "support@ordersmanagement.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a list of orders with optional filters",
                "operationId": "2fd056a400bd274a5b2278e92a8ed010",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter orders by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Filter orders by description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter orders by date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "orders": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "voluptas"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Et cupiditate aut est sed nobis."
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "1978-02-13T00:00:00Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-09-11T07:50:15.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-09-11T07:50:15.000000Z"
                                                    },
                                                    "products": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 56
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "est"
                                                                },
                                                                "price": {
                                                                    "type": "string",
                                                                    "example": "23.43"
                                                                },
                                                                "created_at": {
                                                                    "type": "string",
                                                                    "format": "date-time",
                                                                    "example": "2024-09-11T07:50:15.000000Z"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string",
                                                                    "format": "date-time",
                                                                    "example": "2024-09-11T07:50:15.000000Z"
                                                                },
                                                                "pivot": {
                                                                    "properties": {
                                                                        "order_id": {
                                                                            "type": "integer",
                                                                            "example": 1
                                                                        },
                                                                        "product_id": {
                                                                            "type": "integer",
                                                                            "example": 56
                                                                        },
                                                                        "quantity": {
                                                                            "type": "integer",
                                                                            "example": 4
                                                                        },
                                                                        "created_at": {
                                                                            "type": "string",
                                                                            "format": "date-time",
                                                                            "example": "2024-09-11T07:50:15.000000Z"
                                                                        },
                                                                        "updated_at": {
                                                                            "type": "string",
                                                                            "format": "date-time",
                                                                            "example": "2024-09-11T07:50:15.000000Z"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid parameters"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Store a newly created order in the database along with its products. Request expects an array of products with quantity",
                "operationId": "f92c2f898a00f407e1efa0fd23c93b24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of the order",
                                        "type": "string",
                                        "example": "New Order Name"
                                    },
                                    "description": {
                                        "description": "Description of the order",
                                        "type": "string",
                                        "example": "Description of the new order."
                                    },
                                    "date": {
                                        "description": "The date of the order",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-09-15"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "description": "ID of the product",
                                                    "type": "integer",
                                                    "example": 56
                                                },
                                                "quantity": {
                                                    "description": "Quantity of the product",
                                                    "type": "integer",
                                                    "example": 4
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "order_created": {
                                            "description": "Details of the created order",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "New Order Name"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Description of the new order."
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-15T00:00:00Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-11T07:50:15.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-09-11T07:50:15.000000Z"
                                                },
                                                "products": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 37
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "perferendis"
                                                            },
                                                            "price": {
                                                                "type": "string",
                                                                "example": "10.88"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-09-11T07:50:15.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-09-11T07:50:15.000000Z"
                                                            },
                                                            "pivot": {
                                                                "properties": {
                                                                    "order_id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "product_id": {
                                                                        "type": "integer",
                                                                        "example": 37
                                                                    },
                                                                    "quantity": {
                                                                        "type": "integer",
                                                                        "example": 4
                                                                    },
                                                                    "created_at": {
                                                                        "type": "string",
                                                                        "format": "date-time",
                                                                        "example": "2024-09-11T07:50:15.000000Z"
                                                                    },
                                                                    "updated_at": {
                                                                        "type": "string",
                                                                        "format": "date-time",
                                                                        "example": "2024-09-11T07:50:15.000000Z"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input parameters"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Order creation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order creation failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}